Refactor calculator controller into modular architecture
